requirements.txt
streamlit
python-dotenv
groq
langchain
langchain-community
sentence-transformers
pypdf
chromadb
